Complete: browser/firestore.js
