+++ plugin.xml
@@ -16,14 +16,13 @@
   </engines>
 
   <platform name="browser">
-    <js-module src="www/browser/loader.js" name="loader">
+    <js-module src="www/browser/firestore.js" name="Firestore">
       <clobbers target="Firestore"/>
     </js-module>
-    <js-module src="www/browser/firestore.js" name="Firestore" />
   </platform>
 
   <platform name="android">
-    <js-module src="www/firestore.js" name="Firestore">
+    <js-module src="www/android_ios/firestore.js" name="Firestore">
       <clobbers target="Firestore"/>
     </js-module>
     <config-file target="res/xml/config.xml" parent="/*">
@@ -85,7 +84,7 @@
   </platform>
 
   <platform name="ios">
-    <js-module src="www/firestore.js" name="Firestore">
+    <js-module src="www/android_ios/firestore.js" name="Firestore">
       <clobbers target="Firestore"/>
     </js-module>
     <config-file target="config.xml" parent="/*">
